Add alternative to iconv, fix block size from BLKSSZGET
Appears iconv really is the right approach for this situation, but with
Android's iconv support being unstable there is a problem.  wcstombs()
and mbstowcs() aren't nearly as well-defined, but they do appear to be
working.  Given the broken state of iconv, make use of these.

The block size given by BLKSSZGET appears to always be a 32-bit integer.
Adjust ioctl() call and passing of the value to match.